[TECH] Rename the CLI Command Point namespace to Vertex
