+++ bin/geotools
@@ -41,7 +41,7 @@ use League\Geotools\CLI\Command\Convert;
 use League\Geotools\CLI\Command\Distance;
 use League\Geotools\CLI\Command\Geocoder;
 use League\Geotools\CLI\Command\Geohash;
-use League\Geotools\CLI\Command\Point;
+use League\Geotools\CLI\Command\Vertex;
 use League\Geotools\Geotools;
 
 $geotools = new Geotools;
@@ -60,10 +60,10 @@ $console->add(new Geocoder\Geocode);
 $console->add(new Geocoder\Reverse);
 $console->add(new Geohash\Decode);
 $console->add(new Geohash\Encode);
-$console->add(new Point\Destination);
-$console->add(new Point\FinalBearing);
-$console->add(new Point\FinalCardinal);
-$console->add(new Point\InitialBearing);
-$console->add(new Point\InitialCardinal);
-$console->add(new Point\Middle);
+$console->add(new Vertex\Destination);
+$console->add(new Vertex\FinalBearing);
+$console->add(new Vertex\FinalCardinal);
+$console->add(new Vertex\InitialBearing);
+$console->add(new Vertex\InitialCardinal);
+$console->add(new Vertex\Middle);
 $console->run();
